user_profile/serializer.py
from django.contrib.auth.models import User
from rest_framework import serializers

from .models import Follow

class UserSerializer(serializers.HyperlinkedModelSerializer):
    class Meta:
        model = User
        fields = ['username']

class FollowSerializer(serializers.ModelSerializer):
    class Meta:
        model = Follow
        fields = "__all__"