requirements.txt
aiohttp==3.6.2
appnope==0.1.0
asgiref==3.2.7
async-timeout==3.0.1
attrs==19.3.0
backcall==0.1.0
backoff==1.10.0
bleach==3.1.5
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
cryptography==2.9.2
decorator==4.4.2
defusedxml==0.6.0
Django==3.0.6
django-braces==1.14.0
django-common-helpers==0.9.2
django-cors-headers==3.3.0
django-cron==0.5.1
django-dotenv==1.4.2
django-extensions==2.2.9
django-templated-mail==1.1.1
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
djoser==2.0.3
ecdsa==0.15
entrypoints==0.3
idna==2.9
importlib-metadata==1.6.0
ipykernel==5.3.0
ipython==7.14.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.17.0
Jinja2==2.11.2
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.3
jupyter-console==6.1.0
jupyter-core==4.6.3
MarkupSafe==1.1.1
mistune==0.8.4
multidict==4.7.6
nbconvert==5.6.1
nbformat==5.0.6
notebook==6.0.3
oauthlib==3.1.0
packaging==20.4
pandocfilters==1.4.2
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
prometheus-client==0.7.1
prompt-toolkit==3.0.5
ptyprocess==0.6.0
pyasn1==0.4.8
pycparser==2.20
Pygments==2.6.1
PyJWT==1.7.1
pyparsing==2.4.7
pyrsistent==0.16.0
python-dateutil==2.8.1
python-jose==3.1.0
python3-openid==3.1.0
pytz==2020.1
pyzmq==19.0.1
qtconsole==4.7.4
QtPy==1.9.0
requests==2.23.0
requests-oauthlib==1.3.0
rest-social-auth==4.2.0
rsa==4.0
Send2Trash==1.5.0
six==1.15.0
social-auth-app-django==3.1.0
social-auth-core==3.3.3
spotify==0.10.2
spotipy==2.12.0
sqlparse==0.3.1
terminado==0.8.3
testpath==0.4.4
tornado==6.0.4
traitlets==4.3.3
urllib3==1.25.9
wcwidth==0.1.9
webencodings==0.5.1
widgetsnbextension==3.5.1
yarl==1.4.2
zipp==3.1.0
